Externalize wallet config and seed
